add rgxlog setup file inside spanner_workbench dir
