+++ .github/workflows/test.yml
@@ -10,6 +10,16 @@ jobs:
     steps:
       - name: Checkout code
         uses: actions/checkout@v2
+        
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ env.PYTHON_VERSION }}
+          
+      - name: Install RgxLog
+        run: |
+          python -m pip install --upgrade pip
+          pip install -e ./spanner_workbench
           
       - name: Use nbdev CI Workflow
         uses: fastai/workflows/nbdev-ci@master
